chore: move types outside /interpreter
