+++ packages/hmr/lib/interpreter/index.ts
@@ -1,4 +1,4 @@
-import type { WebSocketMessage, SerializedMessage } from './types';
+import type { SerializedMessage, WebSocketMessage } from '../types';
 
 export default class MessageInterpreter {
   // eslint-disable-next-line @typescript-eslint/no-empty-function
@@ -7,6 +7,7 @@ export default class MessageInterpreter {
   static send(message: WebSocketMessage): SerializedMessage {
     return JSON.stringify(message);
   }
+
   static receive(serializedMessage: SerializedMessage): WebSocketMessage {
     return JSON.parse(serializedMessage);
   }
